changing converter to to_i

+++ RB101/lesson_2/in_lesson_codes/calculator.rb
@@ -10,7 +10,7 @@ def prompt(message)
 end
 
 def valid_number?(num)
-  only_number = num.to_f().to_s() == num
+  only_number = num.to_i().to_s() == num
   starts_with_0 = num.start_with?("0")
   multi_digit = num.length > 1
   only_number && ((!starts_with_0 )|| !multi_digit)
